# [SYCL] Fix WARP_SIZE=16 bug of Intel GPU