- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [ ] Low
  - [x] Medium
  - [ ] High

This PR fixes some bugs of WARP_SIZE=16 for Intel GPU. All warp-related UTs are passed. 
WARP_SIZE=16 has the same output as WARP_SIZE=32 on Intel GPUs.

NOTE: QX_K kernels are specialized for WARP_SIZE=32, so I use a fixed WARP_SIZE for them.

# Performance change
llama-2-7b-chat-hf-q4_0.gguf,  32 in and 32 out, on ARC A770, from 40 tokens/s to 44 tokens/s

## Master Branch
```
 Once upon a time, there existed a little girl, who liked to have adventures. She wanted to go to places and meet new people, have  new experiences and learn new things. She was a curious child, always eager to explore and discover new things.

One day, she found a small door
llama_print_timings:        load time =   14573.13 ms
llama_print_timings:      sample time =       0.53 ms /    32 runs   (    0.02 ms per token, 60952.38 tokens per second)
llama_print_timings: prompt eval time =     217.69 ms /    32 tokens (    6.80 ms per token,   147.00 tokens per second)
llama_print_timings:        eval time =     771.81 ms /    31 runs   (   24.90 ms per token,    40.17 tokens per second)
llama_print_timings:       total time =     994.93 ms /    63 tokens
```
## PR Branch
```
 Once upon a time, there existed a little girl, who liked to have adventures. She wanted to go to places and meet new people, have  new experiences and learn new things. She was a curious child, always eager to explore and discover new things.

One day, she found a small door
llama_print_timings:        load time =   14716.53 ms
llama_print_timings:      sample time =       0.52 ms /    32 runs   (    0.02 ms per token, 61776.06 tokens per second)
llama_print_timings: prompt eval time =     216.68 ms /    32 tokens (    6.77 ms per token,   147.69 tokens per second)
llama_print_timings:        eval time =     698.20 ms /    31 runs   (   22.52 ms per token,    44.40 tokens per second)
llama_print_timings:       total time =     920.69 ms /    63 tokens
```