@qnixsynapse 2024-07-03
Tested iq4_XS, Q4_K_S.

LGTM