@qnixsynapse 2024-07-03
@luoyu-intel  Yes, I am aware. I am testing IQ4 models currently.