@Alcpz 2024-07-03
@joeatodd @OuadiElfarouki Performance of the SYCL branch using an NVIDIA A100 with Q4_K has no regressions.

 | model                          |       size |     params | backend    | ngl |    sm | mmap |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ----: | ---: | ------------: | ---------------: |
| llama 7B Q4_K - Medium         |   3.80 GiB |     6.74 B | SYCL       |  78 |  none |    0 |         pp512 |  2203.66 ± 15.26 |
| llama 13B Q4_K - Medium        |   7.33 GiB |    13.02 B | SYCL       |  78 |  none |    0 |         pp512 |  1720.49 ± 23.80 |
| llama 70B Q4_K - Medium        |  38.58 GiB |    68.98 B | SYCL       |  78 |  none |    0 |         pp512 |    606.95 ± 5.49 |

build: 4887fdce (3293)

| model                          |       size |     params | backend    | ngl |    sm | mmap |          test |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ----: | ---: | ------------: | ---------------: |
| llama 7B Q4_K - Medium         |   3.80 GiB |     6.74 B | SYCL       |  81 |  none |    0 |         tg128 |      5.36 ± 0.00 |
| llama 13B Q4_K - Medium        |   7.33 GiB |    13.02 B | SYCL       |  81 |  none |    0 |         tg128 |      4.27 ± 0.00 |
| llama 70B Q4_K - Medium        |  38.58 GiB |    68.98 B | SYCL       |  81 |  none |    0 |         tg128 |      2.08 ± 0.00 |

build: 4887fdce (3293)